Extract db from main.go
Currently, all the database stuff lives in main.go. Much like the server stuff, it's going to be easier in the long run to have the db stuff encapsulated in its own package. It'll be accessible to main.go AND the future api package. And its nice to have it all in one place, not a super controversial stance.
